@@ -1,12 +1,9 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net8.0</TargetFramework>
 		<AspNetCoreHostingModel>OutOfProcess</AspNetCoreHostingModel>
 	</PropertyGroup>
 	<ItemGroup>
 		<ProjectReference Include="..\Simplify.Web.Postman\Simplify.Web.Postman.csproj" />
-	</ItemGroup>
-	<ItemGroup>
-		<PackageReference Include="Simplify.Web.Json" Version="2.3.*" />
 	</ItemGroup>
 </Project>
