@@ -3,8 +3,6 @@ using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc;
 using Simplify.Web;
 using Simplify.Web.Attributes;
-using Simplify.Web.Json.Responses;
-using TesterApp.ViewModels;
 using TesterApp.ViewModels.Users;
 
 namespace TesterApp.Controllers.Api.v1.Users;
@@ -17,18 +15,18 @@ public class GetMultipleController : Simplify.Web.Controller
 	{
 		var items = new List<UserViewModel>
 		{
-			new UserViewModel
+			new()
 			{
 				UserName = "User 1",
 				CreationTime = DateTime.Now
 			},
-			new UserViewModel
+			new()
 			{
 				UserName = "User 2",
 				CreationTime = DateTime.Now.Subtract(TimeSpan.FromDays(1))
 			}
 		};
 
-		return new Json(items);
+		return Json(items);
 	}
 }
